requirements.txt
torch==2.0.0
ipdb
scipy
transformers==4.35.0
sentencepiece==0.1.97
datasets==2.8.0
tqdm
protobuf==3.20.0
evaluate
wandb
peft
nltk
bitsandbytes
accelerate
wandb
fsspec==2023.9.2
torchvision==0.15.1
scikit-image
open_clip_torch
numexpr
scikit-learn
ruamel.yaml
googletrans==3.1.0a0
rouge-score
